feat(#355): iter 1 - Split commands/fs.rs into sibling module
Split src-tauri/src/commands/fs.rs (429 lines, over the 400-line
architecture rule 23 budget) into a sibling module mirroring the
commands/comments/ pattern:

- commands/fs/mod.rs (220 lines) - canonicalize_path, check_path_exists +
  PathKind, ensure_readable workspace guard, FileStat + stat_file +
  stat_file_inner, update_tree_watched_dirs, plus flat re-exports of every
  submodule item so the existing commands::fs::Foo paths stay valid for
  lib.rs::shared_commands! and integration tests.
- commands/fs/read.rs (118 lines) - MAX_SIZE, TextFileResult, the private
  read_file_capped helper, read_text_file/_inner, read_binary_file/_inner.
- commands/fs/dir.rs (118 lines) - DEFAULT_READ_DIR_LIMIT, ReadDirResult,
  read_dir/_inner with sidecar + sidecar_root filtering.

lib.rs::specta_commands! now points at submodule paths
(commands::fs::dir::read_dir, commands::fs::read::read_text_file,
commands::fs::read::read_binary_file) because the
tauri_specta::collect_commands! macro references compiler-generated
__cmd__/__specta__fn__ siblings of each function that pub use cannot
re-export. Items still callable at commands::fs::Foo via re-export for
every other consumer.

bindings.ts is byte-identical (verified via cargo run --features codegen
+ git diff --exit-code).

Pure code-motion: no behaviour change. Doc citations of commands/fs.rs:NNN
across docs/, src-tauri/src/, and src/ updated to point at the new
submodules with refreshed line numbers.

Refs #355
