+++ docs/architecture.md
@@ -45,7 +45,7 @@ flowchart LR
 2. Every Rust command's TS wrapper is **generated** from `#[specta::specta] #[tauri::command]` annotations into `src/lib/bindings.ts` by `tauri-specta`. The façade in `src/lib/tauri-commands.ts` re-exports each generated wrapper and unwraps the `Result<T, E>` discriminator into `Promise<T>`. Drift between Rust types and the committed `bindings.ts` is mechanically prevented by a CI gate that runs `cargo test --features codegen --test specta_codegen` and asserts `git diff --exit-code src/lib/bindings.ts` (job `bindings-drift` in `.github/workflows/ci.yml`).
 3. Every Rust command is registered in `shared_commands!` in `src-tauri/src/lib.rs:222-262`. Commands are grouped under `src-tauri/src/commands/<feature>.rs` (`src-tauri/src/commands/mod.rs:7-15`): `fs` (incl. `update_tree_watched_dirs` for folder-tree watches), `comments`, `search`, `html`, `launch`, `remote_asset` (bounded HTTPS image proxy — `fetch_remote_asset`; bounds in rule 27 of [`docs/security.md`](security.md)), `config` (`set_author`, `get_author`), `onboarding`, `word_tokens` (`tokenize_words` — UAX#29 word segmentation peer of `compute_anchor_hash`, used by `Word_range` anchors), plus the iter-2 onboarding/platform-integration set (`onboarding` ×1 (`onboarding_state`), `cli_shim` ×3, `default_handler` ×2 — see [`docs/features/installation.md`](features/installation.md)). Path resolution shared by both the GUI and CLI binaries lives in `src-tauri/src/core/paths.rs`.
 
-   *Structured-return chokepoint.* When a single command's caller needs the bytes AND cheap-to-compute metadata (size, line count, …), return a struct that carries both rather than forcing a second IPC round-trip. Canonical example: `read_text_file` returns `TextFileResult { content, size_bytes, line_count }` (`commands/fs.rs:71-109`) so the StatusBar reads from a Zustand-cached `fileMetaByPath` populated by `useFileContent` — neither component issues its own metadata IPC. See rule 2 in [`docs/performance.md`](performance.md).
+   *Structured-return chokepoint.* When a single command's caller needs the bytes AND cheap-to-compute metadata (size, line count, …), return a struct that carries both rather than forcing a second IPC round-trip. Canonical example: `read_text_file` returns `TextFileResult { content, size_bytes, line_count }` (`commands/fs/read.rs:13-110`) so the StatusBar reads from a Zustand-cached `fileMetaByPath` populated by `useFileContent` — neither component issues its own metadata IPC. See rule 2 in [`docs/performance.md`](performance.md).
 4. All frontend logging goes through `src/logger.ts`; no file outside `src/logger.ts` and its test imports from `@tauri-apps/plugin-log`.
 5. Log prefix tags: frontend `[web]`, Rust `[rust]` or a subsystem like `[watcher]`. (`src/logger.ts:9-13`; `watcher.rs:93`.)
 6. `console.log`/`console.info` never appear in production frontend code. Diagnostic logging in watcher hooks goes through `@/logger` (`warn`/`debug`), not `console.*`. (`useFileWatcher.ts:45,57,61`.)
@@ -72,11 +72,11 @@ flowchart LR
 19. Components subscribe to the store with narrow selectors (single-field or `useShallow`), never unfiltered `useStore()`. (`App.tsx:49-63`; `TabBar.tsx:8-10`.) See rule 30 for tier-specific subscription rules.
 20. Comment mutation UI uses `useCommentActions` (`src/lib/vm/use-comment-actions.ts`); components never call low-level `addComment`/`editComment` wrappers. (`CommentThread.tsx:30,113`.)
 21. Comment rendering reads through `useComments` (`src/lib/vm/use-comments.ts`); components never call `getFileComments` directly.
-22. `read_dir` filters out sidecar files (`.review.yaml`, `.review.json`) before returning. (`commands/fs.rs:49-51`.)
+22. `read_dir` filters out sidecar files (`.review.yaml`, `.review.json`) before returning. (`commands/fs/dir.rs:94`.)
 22a. **Viewer perf budget chokepoint** — `src/lib/viewer-budgets.ts` is the single canonical home for all viewer numeric constants (`SIZE_WARN_THRESHOLD`, `MARKDOWN_VISUAL_CAP_BYTES`, `SOURCE_HIGHLIGHT_CHUNK_LINES`, `SOURCE_HIGHLIGHT_IDLE_BUDGET_MS`, `SOURCE_OVERSCAN`, `SOURCE_BASE_LINE_PX`). Inlining a numeric ceiling in a component or hook file is a same-PR debt to extract here so the budget set stays auditable in one place.
 
 ### File-size budgets
-23. Any file >400 lines in `src/components/` or `src-tauri/src/` is a structural smell and must be split. Shared-chokepoint files (`src/store/index.ts`, `src/App.tsx`, `src-tauri/src/lib.rs`) get a 500-line budget. Auto-generated files (`src/lib/bindings.ts`) and test files (`src-tauri/src/**/tests.rs`) are exempt — generated code can't be split, and exhaustive coverage tables grow naturally. **Files currently exceeding their budget** (must be split before this PR's follow-up issue closes): `src-tauri/src/lib.rs` 1061/500 (extract `BINDINGS_HEADER` + `build_specta_builder` + the codegen export block into `specta_export.rs`; also extract `build_window_menu` and `route_args_through_registry` — `menu_event_delivery` was extracted to `src-tauri/src/menu.rs` in PR #341); `src-tauri/src/core/comments.rs` 569/400; `src-tauri/src/commands/comments/badges.rs` 561/400; `src-tauri/src/commands/comments/mod.rs` 533/400; `src-tauri/src/core/types/wire.rs` 474/400; `src-tauri/src/commands/fs.rs` 459/400 (added in iter 1 of #252 for the security-critical `read_file_capped` helper extraction; split scheduled after #252's iter 3 lands to avoid rebase churn); `src-tauri/src/core/matching.rs` 417/400 (17 lines over). All other files in `src/` and `src-tauri/src/` are under budget. Compliance is spot-checked at PR review; a periodic `wc -l` pass refreshes the over-budget list.
+23. Any file >400 lines in `src/components/` or `src-tauri/src/` is a structural smell and must be split. Shared-chokepoint files (`src/store/index.ts`, `src/App.tsx`, `src-tauri/src/lib.rs`) get a 500-line budget. Auto-generated files (`src/lib/bindings.ts`) and test files (`src-tauri/src/**/tests.rs`) are exempt — generated code can't be split, and exhaustive coverage tables grow naturally. **Files currently exceeding their budget** (must be split before this PR's follow-up issue closes): `src-tauri/src/lib.rs` 1061/500 (extract `BINDINGS_HEADER` + `build_specta_builder` + the codegen export block into `specta_export.rs`; also extract `build_window_menu` and `route_args_through_registry` — `menu_event_delivery` was extracted to `src-tauri/src/menu.rs` in PR #341); `src-tauri/src/core/comments.rs` 569/400; `src-tauri/src/commands/comments/badges.rs` 561/400; `src-tauri/src/commands/comments/mod.rs` 533/400; `src-tauri/src/core/types/wire.rs` 474/400; `src-tauri/src/core/matching.rs` 417/400 (17 lines over). All other files in `src/` and `src-tauri/src/` are under budget. (`src-tauri/src/commands/fs.rs` was split into `commands/fs/{mod,read,dir}.rs` in #355.) Compliance is spot-checked at PR review; a periodic `wc -l` pass refreshes the over-budget list.
 
 ### Native menu
 24. Native OS menu events are forwarded as `menu-*` Tauri events handled in `src/hooks/useMenuListeners.ts`, not invoked as commands. Routing lives in `src-tauri/src/menu.rs::menu_event_delivery` + `dispatch_menu_event`, called from `lib.rs::on_menu_event`. **Window-frame and devtools actions are an exception** — they are handled entirely in Rust inside `on_menu_event` (no `menu-*` event, no frontend listener) because they operate on the `WebviewWindow` resolved from the menu-id label and have no UI side-effect that React needs to learn about: `win-minimize`, `win-bring-all`, `new-window`, `toggle-devtools` (latter requires the `devtools` Cargo feature on `tauri` so it works in release builds — see [`docs/features/app-chrome.md`](features/app-chrome.md)).
