+++ docs/security.md
@@ -13,10 +13,10 @@ Canonical for threat-model and safety rules. Cite violations as "violates rule N
 ## Rules
 
 ### File-read bounds
-1. Every Rust command that opens a file enforces a 10 MB hard cap. The cap is a single module-level `const MAX_SIZE: usize = 10 * 1024 * 1024;` shared by both readers. Both `read_text_file_inner` and `read_binary_file_inner` go through a single private helper `read_file_capped` that pre-checks `metadata().len()` on the open handle (fstat — TOCTOU-safe vs. path swap) before allocating, then pairs it with a `File::take(MAX_SIZE + 1)` + post-read length check so special files (`/dev/zero`, FIFOs) reporting `len() == 0` cannot slip through. (`commands/fs.rs:203-225` defines `read_file_capped`: `:212-214` fstat pre-check, `:215-221` bounded `Vec::with_capacity` + `take(MAX_SIZE+1)` read, `:222-224` post-read length check.)
-2. `read_text_file` rejects binaries by scanning the first 512 bytes for NUL, and only succeeds on valid UTF-8. (`commands/fs.rs:250-253` NUL scan; `:257-260` UTF-8 decode — both in `read_text_file_inner`, after `read_file_capped` has bounded the read.)
+1. Every Rust command that opens a file enforces a 10 MB hard cap. The cap is a single module-level `const MAX_SIZE: usize = 10 * 1024 * 1024;` shared by both readers. Both `read_text_file_inner` and `read_binary_file_inner` go through a single private helper `read_file_capped` that pre-checks `metadata().len()` on the open handle (fstat — TOCTOU-safe vs. path swap) before allocating, then pairs it with a `File::take(MAX_SIZE + 1)` + post-read length check so special files (`/dev/zero`, FIFOs) reporting `len() == 0` cannot slip through. (`commands/fs/read.rs:44-67` defines `read_file_capped`: `:53-55` fstat pre-check, `:56-62` bounded `Vec::with_capacity` + `take(MAX_SIZE+1)` read, `:63-65` post-read length check.)
+2. `read_text_file` rejects binaries by scanning the first 512 bytes for NUL, and only succeeds on valid UTF-8. (`commands/fs/read.rs:91-94` NUL scan; `:98-101` UTF-8 decode — both in `read_text_file_inner`, after `read_file_capped` has bounded the read.)
 3. Size pre-check uses `file.metadata()` on the open handle (fstat, not a second path lookup) and is paired with `File::take(MAX+1)` + post-read length check to defend against special files (`/dev/zero`, FIFOs, named pipes, network FS) reporting `len() == 0` while streaming unbounded bytes. Mirrors `core/sidecar.rs::read_capped`. The cap covers `load_sidecar` and `patch_comment`; sidecar YAML further rejects anchors/aliases before parsing (see rule 8).
-4. `read_dir` canonicalizes the requested path and rejects any request whose canonical form differs from the canonicalized input. (`commands/fs.rs:19-32`.)
+4. `read_dir` canonicalizes the requested path and rejects any request whose canonical form differs from the canonicalized input. (`commands/fs/dir.rs:43-55`.)
 
 ### Sidecar atomicity & integrity
 5. Sidecar writes are temp-file + atomic rename; a failed rename cleans up the temp file. (`core/sidecar.rs:91-101`.)
@@ -69,7 +69,7 @@ Canonical for threat-model and safety rules. Cite violations as "violates rule N
 
 - **UI-redress via narrow `style` allow-list (KaTeX).** `style` is permitted on `span`/`math` so KaTeX can lay out math, but only when the element (or an ancestor) is katex-classed (`rehype-katex-style`). The remaining residual risk is a katex-classed wrapper crafted in raw HTML to host an oversized transparent overlay over interactive UI. Defenses in depth: CSP `img-src` blocks remote fetches and there is no `script-src` exposure here; the page chrome lives outside the markdown DOM; React's reconciler keeps event handlers bound to our own components, not user HTML.
 - **No path-origin restriction on mutation commands.** `add_comment`, `edit_comment`, `delete_comment`, `update_comment`, `add_reply`, `get_file_comments` accept any `file_path` string; a confused renderer call could write `<any_path>.review.yaml`. Mitigation: allowlist against open tabs/root.
-- **`check_path_exists` and `read_binary_file` lack the canonicalization guard used by `read_dir`** (`commands/fs.rs:9-15, 96-109`). A symlink could redirect image loads outside the workspace.
+- **`check_path_exists` and `read_binary_file` lack the canonicalization guard used by `read_dir`** (`commands/fs/mod.rs:56-62` `check_path_exists`; `commands/fs/read.rs:115-122` `read_binary_file`). A symlink could redirect image loads outside the workspace.
 - **Sidecar `selected_text` and `text` have no per-field length limit** (`core/types.rs:17-45`); the file-level 10 MB cap (rule 3) bounds total sidecar size, but a single comment can still occupy most of that budget.
 - **Full file paths are logged unredacted** (across `commands/*.rs` `tracing::error!` sites and `watcher.rs:158`). Shared logs leak workspace structure and usernames.
 - **MRSF schema version gate.** `load_sidecar` rejects sidecars with unsupported major versions (> 1) via `reject_unsupported_version`. Minor versions within major 1 are accepted per spec §5.
