--- src-tauri/src/commands/fs.rs
@@ -1,459 +1,0 @@
-//! Filesystem-facing IPC commands: directory listing and file reads.
-
-use super::is_sidecar_file;
-use crate::core::paths::canonicalize_no_verbatim;
-use crate::core::types::DirEntry;
-use crate::mdr_command;
-use std::io::Read;
-
-/// Canonicalize an absolute path to the long form without the Windows `\\?\`
-/// verbatim prefix. Used by the renderer to normalize paths at workspace-open
-/// and tab-open boundaries so that string-equality comparisons against
-/// scanner output (which itself canonicalizes via `dunce`) match. No
-/// workspace guard — this is the very command callers use to obtain the
-/// canonical form before they have a workspace to validate against.
-#[mdr_command]
-pub fn canonicalize_path(path: String) -> Result<String, String> {
-    canonicalize_no_verbatim(std::path::Path::new(&path))
-        .map(|p| p.to_string_lossy().into_owned())
-        .map_err(|e| {
-            tracing::error!("[rust] canonicalize_path error: {}", e);
-            e.to_string()
-        })
-}
-
-/// Outcome of a `check_path_exists` probe. Serialized as a bare lowercase
-/// string ("file" / "dir" / "missing") so the wire shape matches the prior
-/// hand-mirrored TS literal union — but the enum forces specta to emit the
-/// union in `bindings.ts` so the façade can drop its `as` cast and a future
-/// variant addition fails the codegen drift gate.
-#[derive(serde::Serialize, Debug, specta::Type)]
-#[serde(rename_all = "lowercase")]
-pub enum PathKind {
-    File,
-    Dir,
-    Missing,
-}
-
-/// Check if a path exists and whether it is a directory or file.
-#[mdr_command]
-pub fn check_path_exists(path: String) -> PathKind {
-    match std::fs::metadata(&path) {
-        Ok(meta) if meta.is_dir() => PathKind::Dir,
-        Ok(_) => PathKind::File,
-        Err(_) => PathKind::Missing,
-    }
-}
-
-const DEFAULT_READ_DIR_LIMIT: usize = 250;
-
-/// Capped directory listing: entries + total count + overflow flag.
-#[derive(serde::Serialize, Debug, specta::Type)]
-pub struct ReadDirResult {
-    pub entries: Vec<DirEntry>,
-    pub total: usize,
-    pub has_more: bool,
-}
-
-/// Read directory entries, rejecting path traversal.
-/// Returns at most `limit` entries (default 250) with total count so the
-/// frontend can offer a "Show all N items…" affordance.
-/// Hides `.review.yaml`/`.review.json` sidecar files, and also hides the
-/// `sidecar_root` directory when listing a workspace root with an active
-/// redirect (AC10: prevents users from seeing the internal sidecar store).
-#[mdr_command]
-pub fn read_dir(
-    path: String,
-    limit: Option<usize>,
-    show_sidecars: Option<bool>,
-    config_state: tauri::State<'_, crate::watcher::SidecarConfigState>,
-) -> Result<ReadDirResult, String> {
-    read_dir_inner(path, limit, show_sidecars, &config_state)
-}
-
-/// Inner implementation, decoupled from `tauri::State` so unit/integration
-/// tests can construct a plain `SidecarConfigState` and call this directly
-/// without spinning up a full `tauri::App`.
-pub fn read_dir_inner(
-    path: String,
-    limit: Option<usize>,
-    show_sidecars: Option<bool>,
-    config_state: &crate::watcher::SidecarConfigState,
-) -> Result<ReadDirResult, String> {
-    // Canonicalize to resolve symlinks and reject traversal
-    let canonical = canonicalize_no_verbatim(std::path::Path::new(&path)).map_err(|e| {
-        tracing::error!("[rust] command error: {}", e);
-        e.to_string()
-    })?;
-    // Ensure the canonical path matches the requested one (no breakout)
-    let requested = std::path::Path::new(&path);
-    if requested.is_absolute() {
-        let req_canonical = canonicalize_no_verbatim(requested).map_err(|e| e.to_string())?;
-        if req_canonical != canonical {
-            return Err("path traversal not allowed".into());
-        }
-    }
-
-    // Determine if we should hide a sidecar_root directory.
-    // Only applies when we're listing a workspace root that has sidecar_root configured.
-    let hide_dir_name: Option<String> = config_state
-        .resolve_for_file(&canonical)
-        .and_then(|(ws_root, sr)| {
-            if canonical == ws_root {
-                // We're listing the workspace root — hide the first component of sidecar_root
-                sr.and_then(|p| {
-                    p.components()
-                        .next()
-                        .map(|c| c.as_os_str().to_string_lossy().into_owned())
-                })
-            } else {
-                None
-            }
-        });
-
-    let entries = std::fs::read_dir(&canonical).map_err(|e| {
-        tracing::error!("[rust] command error: {}", e);
-        e.to_string()
-    })?;
-
-    let mut result = Vec::new();
-    let show = show_sidecars.unwrap_or(false);
-    for entry in entries {
-        let entry = entry.map_err(|e| {
-            tracing::error!("[rust] command error: {}", e);
-            e.to_string()
-        })?;
-        let meta = entry.metadata().map_err(|e| {
-            tracing::error!("[rust] command error: {}", e);
-            e.to_string()
-        })?;
-        let name = entry.file_name().to_string_lossy().into_owned();
-        // Sidecar file filter — gate on is_file so a directory whose name
-        // ends in `.review.yaml` (legal on every FS) is never mistaken
-        // for a sidecar file.
-        if !show && !meta.is_dir() && is_sidecar_file(&name) {
-            continue;
-        }
-        // The "Show sidecar files in folder pane" toggle controls every
-        // sidecar artifact uniformly: when OFF (default) we hide both
-        // the inline `.review.{yaml,json}` files AND the `sidecar_root`
-        // directory configured by `.mrsf.yaml`. When ON, both surface so
-        // users can browse `.reviews/` and inspect the raw metadata.
-        if !show {
-            if let Some(ref hide) = hide_dir_name {
-                if name == *hide && meta.is_dir() {
-                    continue;
-                }
-            }
-        }
-
-        let path = entry.path().to_string_lossy().into_owned();
-        result.push(DirEntry {
-            name,
-            path,
-            is_dir: meta.is_dir(),
-        });
-    }
-    result.sort_by(|a, b| match (a.is_dir, b.is_dir) {
-        (true, false) => std::cmp::Ordering::Less,
-        (false, true) => std::cmp::Ordering::Greater,
-        _ => a.name.to_lowercase().cmp(&b.name.to_lowercase()),
-    });
-    let total = result.len();
-    let cap = limit.unwrap_or(DEFAULT_READ_DIR_LIMIT);
-    let has_more = total > cap;
-    result.truncate(cap);
-    Ok(ReadDirResult { entries: result, total, has_more })
-}
-
-/// Result of [`read_text_file`]: file content plus cheap-to-compute metadata
-/// (byte size and line count) that the UI surfaces in the status bar without
-/// requiring a second IPC round-trip.
-#[derive(serde::Serialize, Debug, specta::Type)]
-pub struct TextFileResult {
-    pub content: String,
-    pub size_bytes: u64,
-    pub line_count: usize,
-    /// Last-modified time as epoch milliseconds. `None` if the platform/FS
-    /// does not expose mtime or it is before the UNIX epoch. Mirrors the
-    /// `*_ms` epoch convention used by [`FileStat::mtime_ms`]. Surfaced
-    /// here so callers can detect external edits (mtime jumps) without a
-    /// follow-up `stat_file` IPC round-trip.
-    pub mtime_ms: Option<i64>,
-}
-
-/// Shared 10 MB hard cap. Canonical in `docs/security.md` rule 1.
-const MAX_SIZE: usize = 10 * 1024 * 1024;
-
-/// Read a file's bytes with the canonical 10 MB cap pattern.
-///
-/// Combines:
-///   1. fstat pre-check (open-handle `metadata().len()`) — TOCTOU-safe O(1)
-///      reject for multi-GB files vs. path-swap attacks.
-///   2. `Vec::with_capacity(meta.len().min(MAX_SIZE))` — pre-allocate; bounded
-///      by MAX_SIZE so attacker-controlled `meta.len()` cannot OOM.
-///   3. `File::take(MAX_SIZE + 1)` + post-read `bytes.len() > MAX_SIZE` —
-///      defends against special files (`/dev/zero`, FIFOs, network FS) that
-///      report `len() == 0` while streaming unbounded bytes.
-///
-/// Returns the bytes and the open-handle metadata (caller extracts mtime
-/// from the same `open()` so content + mtime cannot be torn).
-///
-/// Mirrors `core::sidecar::read_capped`. See `docs/security.md` rules 1-3.
-fn read_file_capped(path: &str) -> Result<(Vec<u8>, std::fs::Metadata), String> {
-    let file = std::fs::File::open(path).map_err(|e| {
-        tracing::error!("[rust] command error: {}", e);
-        e.to_string()
-    })?;
-    let meta = file.metadata().map_err(|e| {
-        tracing::error!("[rust] command error: {}", e);
-        e.to_string()
-    })?;
-    if meta.len() > MAX_SIZE as u64 {
-        return Err("file_too_large".into());
-    }
-    let mut bytes = Vec::with_capacity(meta.len().min(MAX_SIZE as u64) as usize);
-    file.take(MAX_SIZE as u64 + 1)
-        .read_to_end(&mut bytes)
-        .map_err(|e| {
-            tracing::error!("[rust] command error: {}", e);
-            e.to_string()
-        })?;
-    if bytes.len() > MAX_SIZE {
-        return Err("file_too_large".into());
-    }
-    Ok((bytes, meta))
-}
-
-/// Read a text file, rejecting binary files and files >10 MB. Returns
-/// decoded UTF-8 + `size_bytes` + `line_count` + `mtime_ms`. Workspace-
-/// allowlisted via [`ensure_readable`]; cap + TOCTOU semantics live in
-/// [`read_file_capped`] (see also `docs/security.md` rules 1-3).
-#[mdr_command]
-pub fn read_text_file(
-    path: String,
-    state: tauri::State<'_, crate::watcher::WatcherState>,
-) -> Result<TextFileResult, String> {
-    let canonical = ensure_readable(&path, state.inner())?;
-    read_text_file_inner(canonical.to_string_lossy().into_owned())
-}
-
-/// Inner impl, decoupled from `tauri::State` so tests can exercise pure I/O.
-pub fn read_text_file_inner(path: String) -> Result<TextFileResult, String> {
-    let (bytes, meta) = read_file_capped(&path)?;
-    let mtime_ms = meta
-        .modified()
-        .ok()
-        .and_then(|t| t.duration_since(std::time::UNIX_EPOCH).ok())
-        .map(|d| d.as_millis() as i64);
-
-    // Detect binary by scanning first 512 bytes for null bytes
-    let scan_len = bytes.len().min(512);
-    if bytes[..scan_len].contains(&0u8) {
-        return Err("binary_file".into());
-    }
-
-    let size_bytes = bytes.len() as u64;
-    let content = String::from_utf8(bytes).map_err(|e| {
-        tracing::error!("[rust] command error: {}", e);
-        "binary_file".to_string()
-    })?;
-    let line_count = content.lines().count();
-
-    Ok(TextFileResult {
-        content,
-        size_bytes,
-        line_count,
-        mtime_ms,
-    })
-}
-
-/// Read a binary file, returning base64-encoded content. Rejects files
-/// >10 MB. Workspace-allowlisted via [`ensure_readable`]; cap semantics
-/// live in [`read_file_capped`].
-#[mdr_command]
-pub fn read_binary_file(
-    path: String,
-    state: tauri::State<'_, crate::watcher::WatcherState>,
-) -> Result<String, String> {
-    let canonical = ensure_readable(&path, state.inner())?;
-    read_binary_file_inner(canonical.to_string_lossy().into_owned())
-}
-
-/// Inner impl, no workspace guard. Mirrors [`read_text_file_inner`].
-pub fn read_binary_file_inner(path: String) -> Result<String, String> {
-    let (bytes, _meta) = read_file_capped(&path)?;
-    use base64::Engine;
-    Ok(base64::engine::general_purpose::STANDARD.encode(&bytes))
-}
-
-/// Reject a read request whose canonical path is either outside the watcher
-/// allowlist OR inside a system-locations bucket per
-/// [`crate::core::security::system_locations`]. Used by [`read_text_file`] and
-/// [`read_binary_file`] to close the "no workspace guard" gap called out in
-/// issue #338's review of the existing IPC surface.
-///
-/// Returns the canonical [`PathBuf`] on success so callers perform the actual
-/// I/O against the canonicalized form (defense-in-depth against TOCTOU
-/// symlink swaps between the guard check and the read).
-///
-/// All rejection paths return one of four distinct sentinels so tests (and
-/// the renderer's CommentsPanel filter) can identify which guard fired:
-///
-/// * `"path not in workspace"` — both `is_path_allowed` checks (raw and
-///   post-canonicalize containment). Matched verbatim by
-///   `src/components/comments/CommentsPanel.tsx`.
-/// * `"canonicalize failed"` — `canonicalize_no_verbatim` errored.
-/// * `"system path blocked"` — classify returned `Tier::System` (post
-///   containment, the OS-locations DENY list fired).
-/// * `"path not canonicalizable"` — classify returned a `NonCanonicalErr`.
-///
-/// Workspace-root semantics (issue #338 / iter-1 forward-fix):
-/// `is_path_allowed` is the source of truth for containment — it scans every
-/// window's tree-watched-dirs and watched-paths. Once that has accepted the
-/// path, `classify` is invoked **only** to enforce the system-locations
-/// DENY list (Tier::System). The `workspace_root` argument to `classify`
-/// therefore becomes irrelevant for the Inside/Outside discriminator; we
-/// pass the canonical path itself so the discriminator collapses to
-/// Inside (`canonical.starts_with(canonical)` is always true) and only the
-/// Tier::System branch can reject. Group B's per-window state lands a
-/// proper per-window workspace_root.
-pub fn ensure_readable(
-    path_str: &str,
-    state: &crate::watcher::WatcherState,
-) -> Result<std::path::PathBuf, String> {
-    use crate::core::security::system_locations::{classify, Tier};
-
-    let raw = std::path::Path::new(path_str);
-    // First containment check on the raw path (cheap; matches existing
-    // `stat_file_inner` pattern at line ~292).
-    if !state.is_path_allowed(raw) {
-        tracing::warn!(target: "fs-guard", "[fs-guard] path outside workspace: {}", path_str);
-        return Err("path not in workspace".into());
-    }
-    // Then canonicalize and re-check containment + system-locations.
-    let canonical = canonicalize_no_verbatim(raw)
-        .map_err(|e| {
-            tracing::warn!(target: "fs-guard", "[fs-guard] canonicalize failed for {}: {e}", path_str);
-            "canonicalize failed".to_string()
-        })?;
-    if !state.is_path_allowed(&canonical) {
-        tracing::warn!(target: "fs-guard", "[fs-guard] canonical path outside workspace: {}", canonical.display());
-        return Err("path not in workspace".into());
-    }
-    // `is_path_allowed` has already vetted containment in some window's
-    // tree. We pass `&canonical` as the workspace_root placeholder so the
-    // Inside/Outside discriminator collapses to Inside; only the
-    // Tier::System branch can reject from here. See the doc comment above.
-    match classify(&canonical, &canonical) {
-        Ok(Tier::System { flavor }) => {
-            tracing::warn!(target: "fs-guard", "[fs-guard] system path blocked ({:?}): {}", flavor, canonical.display());
-            Err("system path blocked".into())
-        }
-        Ok(Tier::Inside) | Ok(Tier::Outside) => Ok(canonical),
-        // Defensive branch: per docs/architecture.md rule 11a, every path
-        // reaching `classify(canonical, _)` here has already passed
-        // `canonicalize_no_verbatim` above — so `Tier::classify`'s
-        // `NonCanonicalErr` (rejecting `..`, relative, or verbatim-form
-        // input) is unreachable in practice. Kept as a fail-closed guard
-        // against future refactors that might bypass the canonicalize step.
-        // Per test-expert review iter 1: this branch is intentionally
-        // untested because it is unreachable through the public IPC
-        // contract.
-        Err(e) => {
-            tracing::warn!(target: "fs-guard", "[fs-guard] non-canonical: {} reason={:?}", canonical.display(), e);
-            Err("path not canonicalizable".into())
-        }
-    }
-}
-
-/// Lightweight `stat`: returns just the byte size of a file, with no content
-/// read. Used by viewers (BinaryPlaceholder, TooLargePlaceholder) that need
-/// to display a size without paying the I/O cost of `read_binary_file`. No
-/// 10 MB cap — over-cap files are exactly the case we want to surface.
-///
-/// Workspace-allowlisted: mirrors `commands/system.rs::reveal_in_folder` so a
-/// malicious renderer cannot probe arbitrary paths (e.g. `~/.ssh/id_rsa`)
-/// for existence/size. The path must be inside an open workspace folder or
-/// an open tab.
-#[derive(serde::Serialize, Debug, specta::Type)]
-pub struct FileStat {
-    pub size_bytes: u64,
-    /// Last-modified time as epoch milliseconds. `None` if the platform/FS
-    /// does not expose mtime or it is before the UNIX epoch. Field name
-    /// mirrors the MRSF `*_ms` epoch convention.
-    pub mtime_ms: Option<i64>,
-}
-
-#[mdr_command]
-pub fn stat_file(
-    path: String,
-    state: tauri::State<'_, crate::watcher::WatcherState>,
-) -> Result<FileStat, String> {
-    stat_file_inner(&path, &state)
-}
-
-/// Inner implementation, decoupled from `tauri::State` so unit/integration
-/// tests can construct a plain `WatcherState` and call this directly without
-/// spinning up a full `tauri::App`.
-pub fn stat_file_inner(
-    path: &str,
-    state: &crate::watcher::WatcherState,
-) -> Result<FileStat, String> {
-    if !state.is_path_allowed(std::path::Path::new(path)) {
-        tracing::warn!("[fs] stat_file rejected: path outside workspace");
-        return Err("path not in workspace".into());
-    }
-    let meta = std::fs::metadata(path).map_err(|e| {
-        tracing::error!("[rust] command error: {}", e);
-        e.to_string()
-    })?;
-    let mtime_ms = meta
-        .modified()
-        .ok()
-        .and_then(|t| t.duration_since(std::time::UNIX_EPOCH).ok())
-        .map(|d| d.as_millis() as i64);
-    Ok(FileStat {
-        size_bytes: meta.len(),
-        mtime_ms,
-    })
-}
-
-/// Update the set of directories whose direct children should produce
-/// `folder-changed` events (root + currently-expanded folders in the tree pane).
-///
-/// `root` and every entry in `dirs` are canonicalized internally; callers may
-/// pass any absolute form. Each entry must exist and be a directory, and every
-/// dir must be contained within `root`. At most
-/// [`crate::watcher::MAX_TREE_WATCHED_DIRS`] entries per call.
-///
-/// Also loads the `.mrsf.yaml` config for the workspace root and caches
-/// its `sidecar_root` value in [`SidecarConfigState`] so that subsequent
-/// sidecar reads/writes use the configured path.
-#[mdr_command]
-pub fn update_tree_watched_dirs(
-    window: tauri::Window,
-    root: String,
-    dirs: Vec<String>,
-    state: tauri::State<'_, crate::watcher::WatcherState>,
-    config_state: tauri::State<'_, crate::watcher::SidecarConfigState>,
-) -> Result<(), String> {
-    state.set_tree_watched_dirs(window.label(), root.clone(), dirs).map_err(|e| {
-        tracing::warn!("[rust] update_tree_watched_dirs rejected: {}", e);
-        e
-    })?;
-
-    // Load .mrsf.yaml config for this workspace root.
-    let canonical_root = crate::core::paths::canonicalize_no_verbatim(std::path::Path::new(&root))
-        .map_err(|e| format!("cannot canonicalize root: {e}"))?;
-    let config = crate::core::paths::load_mrsf_config(&canonical_root).unwrap_or_else(|e| {
-        tracing::warn!(
-            "[sidecar-config] failed to load .mrsf.yaml for {}: {e}",
-            root
-        );
-        None
-    });
-    config_state.set_config(canonical_root, config);
-
-    Ok(())
-}
